requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
geopy==2.4.0
flask-sqlalchemy==3.1.1
gunicorn==21.2.0
Werkzeug==2.3.7 